try to create table_helper as a class
